fixed review remarks
